+++ app/webroot/js/collections/todo-collection.js
@@ -1,8 +1,10 @@
 define(function (require) {
-    var TodoModel = require('models/todo-model')
+    var TodoModel = require('models/todo-model');
+    var userID = window.application.loginUser.get('id');
     var TodoCollection = Backbone.Collection.extend({
             url: '/rest-study/todo_lists.json',
             model: TodoModel,
+
 
             parse: function (response) {
                 //コレクションをパース
@@ -11,20 +13,35 @@ define(function (require) {
             },
 
             resetAll: function () {
+
+
                 var eachProcess = function (model) {
-                    model.set('status', '0');
-                    model.save();
+
+                    var owner = model.get('Owner').id;
+                    var assignee = model.get('Assignee').id;
+
+                    if (owner == userID || assignee == userID) {
+                        model.set('status', '0');
+                        model.save();
+                    }
+
                 };
                 this.each(eachProcess);
             },
 
             deleteAll: function () {
 
-                console.log("delete all");
                 var eachProcess = function (model) {
+
+                    var owner = model.get('Owner').id;
+                    
+
+                    if (owner == userID) {
                         model.destroy({
                             wait: true
-                });}
+                        })
+                    }
+                }
                 this.each(eachProcess);
 
             },
